Adjust to changes in recent Leptos 0.5 pre-release versions
Some changes were in Leptos beta2 (change signal value types to
associated types), some in rc1 (logging namespace change), some in rc3
(generate routes is no longer async).

Also run cargo upgrade in the with-server example.

The with-server demo is buggy if you allow location access, it regularly
resets the map and zoom level to go to that location.
The bug was also present when this ran with leptos 0.5.0-beta1